web/requirements.txt
Django==1.4.2
django-tastypie==0.9.11
gevent==0.13.8
greenlet==0.4.0
gunicorn==0.15.0
mimeparse==0.1.3
ordereddict==1.1
python-dateutil==1.5
wsgiref==0.1.2
